kf2rawdat: Fix inaccuracy in converting flux samples to nanoseconds.
The conversion multiplication incurs loss of significance because the
multiplier is a small integer. The resulting error factor was nearly
5%. Instead convert to picoseconds (using a more accurate multiplier
which is ~1000x larger) and then divide down to nanoseconds. Since the
samples are typically 50-100 sample periods, we gain around 6 extra
bits of significance in the result this way, bringing our accuracy to
<0.1% error. This is well within bounds of a physical drive's RPM
accuracy, for example.