libdisk: IPF: Pad the start of the first block on each track.
Because of the way Kryoflux/DTC writes track data to a real floppy
disk, if we start the track with real data (e.g., the 44894489 sync
mark) then the first bit, and hence the entire first block, is lost.

Avoid this by pre-pending a dummy MFM longword 0xAAAAAAAA. This
makes the first block in each track a longword longer than the rest.
For most uses of an IPF, which involve generating convincing MFM
streams, it doesn't really matter too much about the sanity of
metadata such as block lengths.