libdisk: Alternative method for generating weak-bit flux areas.
No flux transitions on write is unreliable; it depends on details
of the read circuit too much.

Instead generate a sequence of bits in the range 3-6us, which should
stray outside the decoder's timing window.

To ensure randomness on replay, jitter samples a little of we read
a single-revolution SCP image.