Make diskread format use average of recent timing measurements
This is important because even the shortest possible read / check
if valid is about 18 cycles, which on an amiga 1000 is almost 2
timer ticks.  This of the order of 20% which means we get extra
bits inserted into the stream.

Averaging them over a few bytes mean that jitter like this is much
less, and we should still be able to detect genuinely fast / slow
sections on a track.

Currently uses a weighting of 25% for each new byte.