Faster diskread
Write CIATALO and ICR direct to memory, then come back and
merge the difference and DSKIDX into a single byte afterwards.
This is slower overall, but faster in the loop where we can
miss bytes from the disk if they are not read in time.

It also requires a buffer 50% bigger.